Change Dictionary Size selection option from 12-30 to 0-40. Adopt the LZMA2 reduced dictionary set which includes 4096 bytes to 4GB-1. Unless you have lots of RAM, most larger dictionary sizes will be reduced by lrzip-next. See open_stream_out function. control.dictSize member changed correctly to u32. Dictionary size limited to 4GB-1 which is 32 bits wide. 0:  2*2^11, 1:  3*2^11 ( 4,096,  6,144) 2:  2*2^12, 3:  3*2^12 ( 8,192, 12,288) 4:  2*2^13  5:  3*2^13 (16,384, 24,576) ... 38: 2*2^30, 39: 3*2^30 (2,147,483,648, 3,221,225,472) 40: 2*2^31 -1 (4,294,967,295 0xFFFFFFFF)

+++ src/include/lrzip_private.h
@@ -441,7 +441,7 @@ struct rzip_control {
 	unsigned filter_flag;		// flag for filters
 	unsigned delta;			// delta flag offset (default 1)
 	unsigned char lzma_properties[5];	// lzma properties, encoded
-	i64 dictSize;			// lzma Dictionary size - set in overhead computation
+	u32 dictSize;			// lzma Dictionary size - set in overhead computation
 	unsigned zpaq_level;		// zpaq level
 	unsigned zpaq_bs;		// zpaq default block size
 	i64 window;
